Open one accordion at a time
As a first attempt without reference, I refactored the code to only allow one accordion to be active at a time. This involved recording the id of the previously active accordion and comparing it to the current accordion being toggled. If it's a match, it gets closed, but if it's different, the previous active accordion is closed while the current accordion is activated. As of now, the code is far from optimal but it gets the job done.